fix: resolve clang-tidy errors and add fuzzing harnesses
Header fixes:
- Add missing SocketBuf_Failed exception declaration to SocketBuf.h
- Add missing SocketCommon_Failed exception declaration to SocketCommon.h
- Fix ARENA_FAILURE macro parenthesization in SocketConfig.h

Code fixes:
- Fix implicit widening conversions in test_arena.c
- Fix implicit widening conversion in fuzz_socketbuf.c
- Add missing default case in fuzz_socketbuf.c switch statement

clang-tidy configuration:
- Suppress false positives from setjmp/longjmp exception handling
- Suppress C idiom patterns (void* casts, widening conversions)
- Document rationale for each suppressed check

New fuzzing harnesses:
- fuzz_arena.c: Arena allocator fuzzing
- fuzz_socketbuf.c: Circular buffer operations
- fuzz_socketbuf_stress.c: Multi-operation buffer stress testing
- fuzz_dns_validate.c: DNS hostname validation
- fuzz_ip_parse.c: IP address parsing
- Include seed corpus data for all fuzzers

Build system:
- Add CMake support for building fuzz targets with libFuzzer

+++ src/test/test_arena.c
@@ -205,7 +205,7 @@ TEST (arena_large_allocation)
   ASSERT_NOT_NULL (arena);
 
   /* Allocate a reasonably large chunk */
-  size_t large_size = 1024 * 1024; /* 1MB */
+  size_t large_size = (size_t)1024 * 1024; /* 1MB */
   void *ptr = ALLOC (arena, large_size);
   ASSERT_NOT_NULL (ptr);
 
@@ -403,7 +403,7 @@ TEST (arena_very_large_allocation)
   ASSERT_NOT_NULL (arena);
 
   /* Allocate close to maximum size */
-  size_t large_size = 10 * 1024 * 1024; /* 10MB */
+  size_t large_size = (size_t)10 * 1024 * 1024; /* 10MB */
   void *ptr = ALLOC (arena, large_size);
   ASSERT_NOT_NULL (ptr);
